+++ src/skaven/WarpLightningCannon.cpp
@@ -76,7 +76,7 @@ int WarpLightningCannon::generateMortalWounds(const Unit *unit)
         bool moreMoreWarpLightning = ((Dice::rollD6() >= 2) || remainingWounds() <= 2);
 
         // Warp Lightning Blast
-        if (distanceTo(m_shootingTarget) <= m_warpLightningBlast.range())
+        if (distanceTo(m_shootingTarget) <= (float)m_warpLightningBlast.range())
         {
             int power = Dice::rollD6();
             Dice::RollResult rollResult;
