+++ src/skaven/WarpfireThrower.cpp
@@ -75,7 +75,7 @@ int WarpfireThrower::generateMortalWounds(const Unit *unit)
         bool moreMoreWarpfire = ((Dice::rollD6() >= 2) || remainingWounds() <= 1);
 
         // Warpfire
-        if (distanceTo(m_shootingTarget) <= m_warpfireThrower.range())
+        if (distanceTo(m_shootingTarget) <= (float)m_warpfireThrower.range())
         {
             int numTargetModels = m_shootingTarget->remainingModels();
             Dice::RollResult rollResult;
